move signing options too
